requirements.txt
chromadb
sentence-transformers
pymupdf
python-docx
beautifulsoup4
python-dotenv
requests
openai